[UWP] Adjust bounds for ContentPage when by itself (#61)
In a scenario where there is a ContentPage by itself, things like labels
ran beneath the StatusBar in landscape mode. This is presumably because
the page was assuming there to be a TitleBar and calculating the bounds
based on that fact; by checking for its visibility and adjusting the
bounds as necessary it allows for correct alignment.