Implemented colored refresh indicator for ListView pull-to-refresh (#2961)
* Implemented colored spinner for ListView pull-to-refresh

* Added back default spinner color

* Renamed away from "spinner" and did some default value magic

Color.Default reinstated, but also added the default black color in the iOS `UpdateRefreshControlColor` method. Seeing this also happens for the seperator color this is something that is necessary. If not, the spinner is invisible.

* Implemented default color on iOS and right property name on Android

For iOS, when the new `RefreshControlColor` has the default color, the `TintColor` for iOS is set to null, causing the color to be the default one.

Forgot to refactor the new name in the Android renderer, whoops

* Reinstated coloring on listview gallry page

* Implemented colored spinner for ListView pull-to-refresh

* Added back default spinner color

* Renamed away from "spinner" and did some default value magic

Color.Default reinstated, but also added the default black color in the iOS `UpdateRefreshControlColor` method. Seeing this also happens for the seperator color this is something that is necessary. If not, the spinner is invisible.

* Implemented default color on iOS and right property name on Android

For iOS, when the new `RefreshControlColor` has the default color, the `TintColor` for iOS is set to null, causing the color to be the default one.

Forgot to refactor the new name in the Android renderer, whoops

* Reinstated coloring on listview gallry page

* Added documentation for new property

* Fixed merge boo-boos