Add ability to set WebView ExecutionMode property (addresses #4720) (#12509)
* To address issue #4720:
Add option for custom renderers to control the execution mode of the WebView control
This is 'opt-in' as by default this commit will not change behavior of existing applciations.  to opt-in, people would need to set the ExecutionMode property in the constructor of their custom WebViewRenderer, like:

  public class MyWebViewRenderer : WebViewRenderer
  {
    public MyWebViewRenderer()
    {
      ExecutionMode = Windows.UI.Xaml.Controls.WebViewExecutionMode.SeparateProcess;
    }
  }

When set as 'SeparateProcess', the memory allocated by the WebView itself is all handled in a sub-process, which ensures the main process never crashes from running out of memory here from doing things like opening and closing youtube every 5 seconds.  This behavior will likely crash when the WebView is in-process due to the huge amounts of memory that is required by that website.

* tabs not spaces

* TemplatedItemsList: Ensure items are unhooked correctly when removing them, whether as individual removals or list resets.  Without this, the native cells do not actually get removed.
CellTableViewCell: use event PropertyChangedEventHandler instead of action (seems more standard)
TextCellRenderer: Correct the event delegate hookup
ListViewRenderer: Actually re-use the UITableViewCell when creating header sections, otherwise we endlessly create new ones, leaving the old ones alive through event handlers

* Revert "TemplatedItemsList: Ensure items are unhooked correctly when removing them, whether as individual removals or list resets.  Without this, the native cells do not actually get removed."

This reverts commit 7da9ffb0c9d8138bf93fd179ded4e92ff4ee601c.

* Make the custom renderer only apply to the test for this issue, as it seems the SeparateProcess affects access to cookies, breaking other tests

* Move additional classes inside test class to decrease namespace cluttering.

* - cleanup and rebase

* - add instructions

* - move WebViewExecutionMode to platform specific

* - fix up UI Test

* - clean up tests

* - fix tabs

* - fix formatting

* - fix teardown call

* - fix async ui test quirk on uwp

Co-authored-by: Shane Neuville <shneuvil@microsoft.com>